requirements.txt
# Core dependencies
requests>=2.31.0
beautifulsoup4>=4.12.0
pandas>=2.0.0
python-dotenv>=1.0.0
lxml>=4.9.0

# Web scraping
selenium>=4.15.0
setuptools>=68.0.0

# Data analysis and visualization
matplotlib>=3.7.0
seaborn>=0.12.0
openpyxl>=3.1.0
plotly>=5.17.0
kaleido>=1.0.0

# Optional: For enhanced analytics
numpy>=1.24.0